fix: сhange extention js to ts
